+++ must-dash/src/composables.rs
@@ -1,1 +1,2 @@
 pub mod formatting;
+pub mod stats;
