+++ must-hop/src/node/mesh_router.rs
@@ -110,7 +110,7 @@ where
     ) -> Result<Vec<MHPacket<SIZE>, LEN>, MeshRouterError<Node::Error>> {
         // Heartbeats, only for GW
         if let Some(heartbeat_pkt) = self.routing_policy.check_heartbeat(&mut self.manager)? {
-            trace!("SENDING OUT HEARTBEAT!!");
+            // trace!("SENDING OUT HEARTBEAT!!");
             // TODO: Make this into a flag for the mac policy, meaning it will send the correct
             // slot together with a heartbeat
             self.mac_policy.tx_heartbeat(heartbeat_pkt);
