--- must-hop/src/node/time_sync.d2
@@ -1,169 +1,0 @@
-direction: right
-vars: {
-  d2-config: {
-    layout-engine: elk
-  }
-}
-
-classes: {
-  block: {
-    shape: rectangle
-    style: {
-      fill: "#FFFFFF"
-      stroke: "#000"
-    }
-  }
-  sum: {
-    shape: circle
-    width: 60
-    height: 60
-    style: {
-      fill: transparent
-      stroke: "#333"
-      stroke-width: 2
-    }
-  }
-}
-
-# --- Inputs (Breaking the massive physical loop) ---
-HB_in: "Heartbeat Packet (From GW)" {
-  shape: step
-  style.fill: "#e0f7fa"
-}
-
-time_stamp: "Time stamp calculation" {
-  TG: |latex T_G[k] | {class: block}
-  TG_with_delay: |latex T_{G} + NW_{delay} | {class: sum}
-  delay: |latex z^{-1} | {class: block}
-  sums: |latex \Delta \text{gw} | {class: sum}
-
-  # TG branches out
-  TG -> TG_with_delay: {
-    target-arrowhead.label: "+"
-  }
-
-  # TG_with_delay branches out
-  TG_with_delay -> delay: |latex T_N[k] |
-  TG_with_delay -> sums: |latex T_N[k] |
-
-  delay -> sums: {
-    target-arrowhead.label: "-"
-    label: |latex T_N[k-1] |
-  }
-}
-
-network_delay: "Network delay calculation" {
-  DU: |latex \Delta up | {class: block}
-  THW: |latex t_{hw}[k] | {class: block}
-  TH_delay: |latex z^{-1} | {class: block}
-  thw_sums: |latex \Sigma | {class: sum}
-  dup_sum: |latex \Sigma | {class: sum}
-  delta_sum: |latex \Sigma | {class: sum}
-  nw_delay_sum: |latex \Sigma | {class: sum}
-  factor: |latex \frac{1}{2} | {class: sum}
-
-  THW -> TH_delay
-  THW -> thw_sums: {
-    target-arrowhead.label: "+"
-  }
-  TH_delay -> thw_sums: {
-    label: |latex t_{hw}[k-1] |
-    target-arrowhead.label: "-"
-  }
-
-  thw_sums -> dup_sum: {
-    target-arrowhead.label: "+"
-    label: |latex \Delta \text{local} |
-  }
-
-  dup_sum -> delta_sum: {
-    target-arrowhead.label: "+"
-  }
-
-  delta_sum -> nw_delay_sum: {
-    target-arrowhead.label: "+"
-    label: |latex \Delta down |
-  }
-
-  DU -> nw_delay_sum: {
-    target-arrowhead.label: "+"
-    label: |latex \Delta up |
-  }
-
-  nw_delay_sum -> factor
-}
-
-skew: "Calculation of skew" {
-  kp: |latex K_p | {class: block}
-  SK_sum: |latex \Sigma | {class: sum}
-  sk_delay: |latex z^{-1} | {class: block}
-
-  kp -> SK_sum
-
-  SK_sum -> sk_delay: |latex S[k] |
-
-  sk_delay -> SK_sum: {
-    target-arrowhead.label: "+"
-    label: |latex S[k-1] |
-  }
-}
-
-# --- External Variables and Error Calc ---
-error_sum: |latex \Sigma | {class: sum}
-
-time_stamp.sums -> error_sum: {
-  target-arrowhead.label: "+"
-  label: |latex \Delta gw |
-}
-network_delay.thw_sums -> error_sum: {
-  target-arrowhead.label: "-"
-}
-
-error_sum -> skew.kp: {
-  label: "error"
-}
-
-now: "now()" {class: block}
-now_sum: |latex \Sigma | {class: sum}
-now_skewed: |latex \times | {class: sum}
-hb_sum: |latex \Sigma | {class: sum}
-
-now -> now_sum: {
-  target-arrowhead.label: "+"
-}
-network_delay.THW -> now_sum: {
-  target-arrowhead.label: "-"
-}
-
-now_sum -> now_skewed
-skew.SK_sum -> now_skewed
-
-now_skewed -> hb_sum: {
-  target-arrowhead.label: "+"
-  label: "Time since last heartbeat"
-}
-time_stamp.TG_with_delay -> hb_sum: {
-  target-arrowhead.label: "+"
-}
-
-# --- Outputs (Breaking the massive physical loop) ---
-GW_out: "GW Broadcast" {
-  shape: step
-  style.fill: "#e0f7fa"
-}
-
-hb_sum -> GW_out: "Node's heartbeat broadcast"
-
-# --- Inter-container routes ---
-HB_in -> time_stamp.TG
-HB_in -> network_delay.DU
-
-network_delay.factor -> time_stamp.TG_with_delay
-
-time_stamp.delay -> network_delay.dup_sum: {
-  target-arrowhead.label: "+"
-}
-
-time_stamp.TG -> network_delay.delta_sum: {
-  target-arrowhead.label: "-"
-}
